Added meshes
Editor:
    * Console Widget, display logs and futurely will handle commands;
    * Work on Components Displayer for displaying mesh properties;

Engine:
    * Scene Rendering is working;
    * Performing CPU-Timed profiling of the initialization and run-loop;
    * Changed how Timestep work, moving to a simpler approach for now;

Renderer:
    * Implemented GLTF parsing for loading meshes;
    * Interfaced Texture2D, TextureCubemap and Mesh in order to facilitate multiple rendering backends in the future;

+++ Engine/Core/Scene.h
@@ -40,7 +40,7 @@ namespace Cosmos::Engine
 	public:
 
 		// updates the scene logic
-		void OnUpdate(double timestep);
+		void OnUpdate(float timestep);
 
 		// draws the scene objects
 		void OnRender();
